fixing the test cases.
